updates 12/14
