+++ demo-zustand-todo/package.json
@@ -10,7 +10,8 @@
     "install": "npm install ../package",
     "update": " npm run uninstall && npm run install && npm run start",
     "buildPackage": "tsc",
-    "switchToPackage": "cd ../package"
+    "switchToPackage": "cd ../package",
+    "symlink": "npm --prefix ../package run build && npm pack ../package && npm uninstall chromogen && npm install ./chromogen-4.0.4.tgz && npm start"
   },
   "keywords": [
     "react",
@@ -69,7 +70,7 @@
     "@mui/icons-material": "^5.10.6",
     "@mui/material": "^5.10.6",
     "babel-jest": "^26.3.0",
-    "chromogen": "file:../package",
+    "chromogen": "file:chromogen-4.0.4.tgz",
     "file-loader": "^6.2.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
