+++ package/index.ts
@@ -1,6 +1,6 @@
 /* eslint-disable */
 import { atom, selector, atomFamily, selectorFamily } from './recoil_generator/src/api/api';
-import { ChromogenZustandObserver } from './zustand_generator/src/component/ChromogenZustandObserver';
+import { ChromogenZustandObserver } from './zustand_generator/src/component/panel';
 import { ChromogenObserver } from './recoil_generator/src/component/ChromogenObserver';
 import { chromogenZustandMiddleware } from './zustand_generator/src/api/api';
